trace : bug fixes + fixed incomplete export after parent did finish + fixed inproper handling of BAR accesses not ending in device + fixed incorrect service name issue

+++ trace/include/events/events.h
@@ -451,7 +451,7 @@ class HostMmioOp : public HostAddrSizeOp {
  public:
   void Display(std::ostream &out) override;
 
-  uint64_t GetBar() const;
+  int GetBar() const;
 
   uint64_t GetOffset() const;
